break out query execution related logic
