+++ public/app/plugins/datasource/cloudwatch/__mocks__/CloudWatchDataSource.ts
@@ -1,62 +1,93 @@
 import { of } from 'rxjs';
 
-import { dateTime } from '@grafana/data';
-import { setBackendSrv } from '@grafana/runtime';
+import {
+  DataSourceInstanceSettings,
+  DataSourcePluginMeta,
+  PluginMetaInfo,
+  PluginType,
+  VariableHide,
+} from '@grafana/data';
+import { getBackendSrv, setBackendSrv } from '@grafana/runtime';
+import { getTimeSrv } from 'app/features/dashboard/services/TimeSrv';
 import { TemplateSrv } from 'app/features/templating/template_srv';
 import { initialCustomVariableModelState } from 'app/features/variables/custom/reducer';
 import { CustomVariableModel } from 'app/features/variables/types';
 
-import { TemplateSrvMock } from '../../../../features/templating/template_srv.mock';
 import { CloudWatchDatasource } from '../datasource';
+import { CloudWatchJsonData } from '../types';
+
+export function setupMockedTemplateService(variables: CustomVariableModel[]) {
+  const templateService = new TemplateSrv();
+  templateService.init(variables);
+  templateService.getVariables = jest.fn().mockReturnValue(variables);
+  return templateService;
+}
+
+const info: PluginMetaInfo = {
+  author: {
+    name: '',
+    url: undefined,
+  },
+  description: '',
+  links: [],
+  logos: {
+    large: '',
+    small: '',
+  },
+  screenshots: [],
+  updated: '',
+  version: '',
+};
+
+export const meta: DataSourcePluginMeta<CloudWatchJsonData> = {
+  id: '',
+  name: '',
+  type: PluginType.datasource,
+  info,
+  module: '',
+  baseUrl: '',
+};
+
+export const CloudWatchSettings: DataSourceInstanceSettings<CloudWatchJsonData> = {
+  jsonData: { defaultRegion: 'us-west-1', tracingDatasourceUid: 'xray' },
+  id: 0,
+  uid: '',
+  type: '',
+  name: 'CloudWatch Test Datasource',
+  meta,
+  readOnly: false,
+  access: 'direct',
+};
 
 export function setupMockedDataSource({
-  data = [],
   variables,
   mockGetVariableName = true,
-}: { data?: any; variables?: any; mockGetVariableName?: boolean } = {}) {
-  let templateService = new TemplateSrvMock({
-    region: 'templatedRegion',
-    fields: 'templatedField',
-    group: 'templatedGroup',
-  }) as any;
+}: {
+  variables?: CustomVariableModel[];
+  mockGetVariableName?: boolean;
+} = {}) {
+  let templateService = new TemplateSrv();
   if (variables) {
-    templateService = new TemplateSrv();
-    templateService.init(variables);
-    templateService.getVariables = jest.fn().mockReturnValue(variables);
+    templateService = setupMockedTemplateService(variables);
     if (mockGetVariableName) {
       templateService.getVariableName = (name: string) => name;
     }
   }
 
-  const datasource = new CloudWatchDatasource(
-    {
-      jsonData: { defaultRegion: 'us-west-1', tracingDatasourceUid: 'xray' },
-    } as any,
-    templateService,
-    {
-      timeRange() {
-        const time = dateTime('2021-01-01T01:00:00Z');
-        const range = {
-          from: time.subtract(6, 'hour'),
-          to: time,
-        };
-
-        return {
-          ...range,
-          raw: range,
-        };
-      },
-    } as any
-  );
+  const timeSrv = getTimeSrv();
+  const datasource = new CloudWatchDatasource(CloudWatchSettings, templateService, timeSrv);
   datasource.getVariables = () => ['test'];
 
   datasource.getNamespaces = jest.fn().mockResolvedValue([]);
   datasource.getRegions = jest.fn().mockResolvedValue([]);
   datasource.defaultLogGroups = [];
-  const fetchMock = jest.fn().mockReturnValue(of({ data }));
-  setBackendSrv({ fetch: fetchMock } as any);
+  const fetchMock = jest.fn().mockReturnValue(of({}));
+  setBackendSrv({
+    ...getBackendSrv(),
+    fetch: fetchMock,
+  });
 
-  return { datasource, fetchMock, templateService };
+  return { datasource, fetchMock, templateService, timeSrv };
 }
 
 export const metricVariable: CustomVariableModel = {
@@ -180,7 +211,7 @@ export const regionVariable: CustomVariableModel = {
   multi: false,
 };
 
-export const expressionVariable: CustomVariableModel = {
+export const fieldsVariable: CustomVariableModel = {
   ...initialCustomVariableModelState,
   id: 'fields',
   name: 'fields',
@@ -191,4 +222,18 @@ export const expressionVariable: CustomVariableModel = {
   },
   options: [{ value: 'templatedField', text: 'templatedField', selected: true }],
   multi: false,
+};
+
+export const periodIntervalVariable: CustomVariableModel = {
+  ...initialCustomVariableModelState,
+  id: 'period',
+  name: 'period',
+  index: 0,
+  current: { value: '10m', text: '10m', selected: true },
+  options: [{ value: '10m', text: '10m', selected: true }],
+  multi: false,
+  includeAll: false,
+  query: '',
+  hide: VariableHide.dontHide,
+  type: 'custom',
 };
