+++ public/app/plugins/datasource/cloudwatch/__mocks__/logsTestContext.ts
@@ -1,0 +1,40 @@
+import { of } from 'rxjs';
+
+import { DataFrame, dataFrameToJSON, MutableDataFrame, ArrayVector } from '@grafana/data';
+
+import { CloudWatchLogsQueryStatus } from '../types';
+
+import { setupMockedDataSource } from './CloudWatchDataSource';
+
+export function setupForLogs() {
+  function envelope(frame: DataFrame) {
+    return { data: { results: { a: { refId: 'a', frames: [dataFrameToJSON(frame)] } } } };
+  }
+
+  const { datasource, fetchMock, timeSrv } = setupMockedDataSource();
+
+  const startQueryFrame = new MutableDataFrame({ fields: [{ name: 'queryId', values: ['queryid'] }] });
+  fetchMock.mockReturnValueOnce(of(envelope(startQueryFrame)));
+
+  const logsFrame = new MutableDataFrame({
+    fields: [
+      {
+        name: '@message',
+        values: new ArrayVector(['something']),
+      },
+      {
+        name: '@timestamp',
+        values: new ArrayVector([1]),
+      },
+      {
+        name: '@xrayTraceId',
+        values: new ArrayVector(['1-613f0d6b-3e7cb34375b60662359611bd']),
+      },
+    ],
+    meta: { custom: { Status: CloudWatchLogsQueryStatus.Complete } },
+  });
+
+  fetchMock.mockReturnValueOnce(of(envelope(logsFrame)));
+
+  return { datasource, fetchMock, timeSrv };
+}
