+++ public/app/plugins/datasource/cloudwatch/components/ConfigEditor.tsx
@@ -53,6 +53,9 @@ export const ConfigEditor: FC<Props> = (props: Props) => {
       .put(`/api/datasources/${options.id}`, options)
       .then((result: { datasource: any }) => {
         updateDatasourcePluginOption(props, 'version', result.datasource.version);
+      })
+      .catch((err) => {
+        console.log(err);
       });
     setSaved(true);
   };
