+++ public/app/plugins/datasource/cloudwatch/query-runner/CloudWatchAnnotationQueryRunner.ts
@@ -1,0 +1,44 @@
+import { map, Observable } from 'rxjs';
+
+import { DataFrame, DataQueryRequest, DataQueryResponse, DataSourceInstanceSettings } from '@grafana/data';
+import { toDataQueryResponse } from '@grafana/runtime';
+import { TemplateSrv } from 'app/features/templating/template_srv';
+
+// import { CloudWatchDataSource } from './CloudWatchDataSource';
+import { CloudWatchAnnotationQuery, CloudWatchJsonData, CloudWatchQuery } from '../types';
+
+import { CloudWatchQueryRunner } from './CloudWatchQueryRunner';
+
+export class CloudWatchAnnotationQueryRunner extends CloudWatchQueryRunner {
+  constructor(instanceSettings: DataSourceInstanceSettings<CloudWatchJsonData>, templateSrv: TemplateSrv) {
+    super(instanceSettings, templateSrv);
+  }
+
+  handleAnnotationQuery(
+    queries: CloudWatchAnnotationQuery[],
+    options: DataQueryRequest<CloudWatchQuery>
+  ): Observable<DataQueryResponse> {
+    return this.awsRequest(this.dsQueryEndpoint, {
+      from: options.range.from.valueOf().toString(),
+      to: options.range.to.valueOf().toString(),
+      queries: queries.map((query) => ({
+        ...query,
+        statistic: this.templateSrv.replace(query.statistic),
+        region: this.templateSrv.replace(this.getActualRegion(query.region)),
+        namespace: this.templateSrv.replace(query.namespace),
+        metricName: this.templateSrv.replace(query.metricName),
+        dimensions: this.convertDimensionFormat(query.dimensions ?? {}, {}),
+        period: query.period ?? '',
+        actionPrefix: query.actionPrefix ?? '',
+        alarmNamePrefix: query.alarmNamePrefix ?? '',
+        type: 'annotationQuery',
+        datasource: this.ref,
+      })),
+    }).pipe(
+      map((r) => {
+        const frames = toDataQueryResponse({ data: r }).data as DataFrame[];
+        return { data: frames };
+      })
+    );
+  }
+}
