+++ public/app/plugins/datasource/cloudwatch/types.ts
@@ -1,5 +1,5 @@
 import { AwsAuthDataSourceJsonData, AwsAuthDataSourceSecureJsonData } from '@grafana/aws-sdk';
-import { DataQuery, DataSourceRef, SelectableValue } from '@grafana/data';
+import { DataFrame, DataQuery, DataSourceRef, SelectableValue } from '@grafana/data';
 
 import {
   QueryEditorArrayExpression,
@@ -244,6 +244,7 @@ export interface TSDBQueryResult<T = any> {
   refId: string;
   series: TSDBTimeSeries[];
   tables: Array<TSDBTable<T>>;
+  frames: DataFrame[];
 
   error?: string;
   meta?: any;
@@ -252,6 +253,15 @@ export interface TSDBQueryResult<T = any> {
 export interface TSDBTable<T = any> {
   columns: Array<{ text: string }>;
   rows: T[];
+}
+
+export interface DataQueryError<CloudWatchMetricsQuery> {
+  data?: {
+    message?: string;
+    error?: string;
+    results: Record<string, TSDBQueryResult<CloudWatchMetricsQuery>>;
+  };
+  message?: string;
 }
 
 export interface TSDBTimeSeries {
